requirements.txt
selenium~=4.20.0
requests~=2.31.0
pytest~=8.2.0
allure-pytest~=2.13.5